// frontend-tsx/src/_state/actions-types/questions/types.tsx
export enum QuestionActionType {
    CREATE_QUESTION_SUCCESS = 'CREATE_QUESTION_SUCCESS',
    CREATE_QUESTION_FAIL = 'CREATE_QUESTION_FAIL',
    DELETE_QUESTION_SUCCESS = 'DELETE_QUESTION_SUCCESS',
    DELETE_QUESTION_FAIL = 'DELETE_QUESTION_FAIL',
    START_QUESTION_SUCCESS = 'START_QUESTION_SUCCESS',
    START_QUESTION_FAIL = 'START_QUESTION_FAIL',
    CURRENT_QUESTION_SUCCESS = 'CURRENT_QUESTION_SUCCESS',
    CURRENT_QUESTION_FAIL = 'CURRENT_QUESTION_FAIL',
    ROOM_QUESTION_SUCCESS = 'ROOM_QUESTION_SUCCESS',
    ROOM_QUESTION_FAIL = 'ROOM_QUESTION_FAIL',
    NR_OF_QUESTIONS_SUCCESS = "NR_OF_QUESTIONS_SUCCESS",
    NR_OF_QUESTIONS_FAIL = "NR_OF_QUESTIONS_FAIL",
    FIRST_QUESTION_SUCCESS = "FIRST_QUESTION_SUCCESS",
    FIRST_QUESTION_FAIL = "FIRST_QUESTION_FAIL",

    ANSWER_QUESTION_SUCCESS = "ANSWER_QUESTION_SUCCESS",
    ANSWER_QUESTION_FAIL = "ANSWER_QUESTION_FAIL",
}